requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
Pillow==10.0.1
google-genai
python-multipart==0.0.6